Better formatting

+++ src/com/jschiff/math/fpprecision/PrecisionPair.java
@@ -13,6 +13,6 @@ public class PrecisionPair {
 
     @Override
     public String toString() {
-        return String.format("%.1f, %.1f, %d", low, high, numberOfValuesBetween);
+        return String.format("%d, %d, %d", (int)low, (int)high, numberOfValuesBetween);
     }
 }
